+++ LinearizedTrackFit/interface/BuildTestFunctions.h
@@ -38,7 +38,7 @@ void transformVariables(const std::vector<double> & vars, const std::vector<int>
                         const T & variablesTransformations, std::vector<double> & transformedVars,
                         const double & genChargeOverPt, const double & genCotTheta, const double & genZ0)
 {
-  for (int i=0; i<vars.size()/3; ++i) {
+  for (size_t i=0; i<vars.size()/3; ++i) {
     for (auto v : variablesTransformations) {
       transformedVars.push_back((*v)(i, vars, uniqueLayers, genChargeOverPt, genCotTheta, genZ0));
     }
